Phase 3M: V1-Runtime-PoC bestanden (VGUI2 unter Xash).
Vendorte vgui_controls + Xash-Surface/Input/System, FreeType-Glyphen,
interaktiver PoC-Nachweis und Living Docs; kein Phase-3-Tag.

+++ client/dep/NclNitroApi/dep/ncl-hl1-source-sdk/tier1/KeyValues.cpp
@@ -2156,7 +2156,8 @@ bool KeyValues::WriteAsBinary( CUtlBuffer &buffer )
 			}
 		case TYPE_PTR:
 			{
-				buffer.PutUnsignedInt( (int)dat->m_pValue );
+				// CS Retro 64-Bit: store pointer as 64-bit in buffer when available.
+				buffer.PutUnsignedInt( (unsigned int)(uintptr_t)dat->m_pValue );
 			}
 
 		default:
